resnet101-image-labels/requirements.txt
blinker==1.8.2
click==8.1.7
filelock==3.13.1
Flask==3.0.3
fsspec==2024.2.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
pillow==10.2.0
setuptools==70.0.0
sympy==1.13.1
torch==2.5.0+cpu
torchaudio==2.5.0+cpu
torchvision==0.20.0+cpu
typing_extensions==4.9.0
Werkzeug==3.0.6
